Production: Remove all files after execution

+++ Util/Del.py
@@ -22,10 +22,10 @@ def delete(source_file, binary_file, input_file, output_file, err_file):
     try:
         os.remove(source_file)
         os.remove(binary_file)
-        os.remove(input_file)
+        if input_file:
+            os.remove(input_file)
         os.remove(output_file)
         os.remove(err_file)
         return 0
-    except OSError as e:
-        print(e)
+    except OSError:
         return -1
